etl/sqlite_to_postgres/requirements.txt
flake8==4.0
python-dotenv==0.20.0
psycopg2-binary==2.9.3